Refactor netmiko code
Move functions that will be used by other netmiko clients (libvirt)
into utils/netmiko, retaining just the core functionality in
cli/netmiko
